Bump node to minimal supported version
This action previously requested node-16 which is unsupported and was
automatically bumped to the minimal supported version node-20. This
triggered a warning for users:

```
The following actions use a deprecated Node.js version and will be forced to run on node20: boa-dev/criterion-compare-action@v3.  For more info: https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```